refactor: rename libs to lib and create utils folder
